+++ python/src/connectors/__init__.py
@@ -1,0 +1,2 @@
+from .azure_chat_completion import RESPONSE_OBJECT_KEY, AzureChatCompletion
+from .open_ai_chat_history import OpenAIChatHistory
