+++ python/src/connectors/azure_chat_completion.py
