+++ python/src/connectors/openai_chat_history.py
