+++ python/src/semantic_functions/__init__.py
@@ -1,5 +1,5 @@
-from .native_function import NativeFunction
+from ..orchestration.native_function import NativeFunction
+from ..orchestration.sk_function import SKFunction
 from .semantic_function import SemanticFunction
-from .sk_function import SKFunction
 from .sk_function_decorator import sk_function
 from .sk_function_parameter_decorator import sk_function_parameter
