+++ python/src/template_engine/__init__.py
@@ -1,0 +1,1 @@
+from .handlebars_prompt_template_handler import HandleBarsPromptTemplateHandler
