+++ python/src/template_engine/handlebars_prompt_template_handler.py
